% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bericht.R
\name{bericht}
\alias{bericht}
\title{Berichte erstellen}
\usage{
bericht(
  continent = "all",
  min_year = 1952,
  max_year = 2007,
  Land1 = "Germany",
  Land2 = "Turkey",
  Dateiname = "Gapminder-Bericht.html"
)
}
\arguments{
\item{continent}{Kontinent}

\item{min_year}{Untere Schwelle: erstes Jahr, das im Bericht berücksichtigt wird}

\item{max_year}{Obere Schwelle: letztes Jahr, das im Bericht berücksichtigt wird}

\item{Land1}{Erstes Land für die Grafik mit dem Ländervergleich}

\item{Land2}{Zweites Land für die Grafik mit dem Ländervergleich}

\item{Dateiname}{Dateiname als Text, z. B. "Gapminder-Bericht.html" (mit Anführungszeichen; Dateiendung: .html)}
}
\value{
Bericht im Word-Format
}
\description{
Berichte erstellen im HTML-Format.
Achtung: Evtl. im aktuellen Arbeitsverzeichnis vorhandener Bericht wird überschrieben!
Empfehlung: Dateinamen angeben!
}
\examples{
bericht()
bericht(continent = "Europe", Land1 = "Iceland", Land2 = "Turkey")
bericht(continent = "Europe", min_year = 1982, max_year = 2007,
 Land1 = "Iceland", Land2 = "Turkey", Dateiname = "Europa.html")
}
